+++ src/errors.js
@@ -1,15 +1,32 @@
 const lodash = require("lodash");
+const { Logger } = require("bunyan")
 
 /**
  * Error class to capture different types of request errors
  */
 class RequestError extends Error {
+    requestError;       // Initial Axios error
+    status;             // Request status code
+
+    /**
+     * Constructor for RequestError
+     * @param {string} name Name of error.
+     * @param {AxiosError} reqError Request error to be used.
+     */
     constructor(name, reqError) {
         super(name);
         // extend stack trace to not lose the existing trace
+        this.requestError = reqError;
+        this.status = reqError.status;
         this.stack += reqError.stack.substring(reqError.stack.indexOf("\n"));
     }
 
+    /**
+     * Make a HaystackError, GraphQLError where applicable. If the given reqError is not determined to be either a
+     * HaystackError or GraphQL error, the original error is returned.
+     * @param {AxiosError | Error} reqError Error to be parsed.
+     * @param {Logger} logger A Bunyan logging instance to log the creation of a new Error instance.
+     */
     static make(reqError, logger) {
         if (lodash.has(reqError, "response.data")) {
             const { data } = reqError.response;
@@ -40,6 +57,12 @@ class RequestError extends Error {
  * Haystack type error (e.g. HisWrite, CreateRec, HisRead, etc.)
  */
 class HaystackError extends RequestError {
+
+    /**
+     * Constructor for HaystackError
+     * @param {string} name Name of error.
+     * @param {AxiosError} reqError Request error to be used.
+     */
     constructor(name, reqError) {
         super(name, reqError);
     }
@@ -49,6 +72,12 @@ class HaystackError extends RequestError {
  * GraphQL type error due to syntax issues
  */
 class GraphQLError extends RequestError {
+
+    /**
+     * Constructor for GraphQLError
+     * @param {string} name Name of error.
+     * @param {AxiosError} reqError Request error to be used.
+     */
     constructor(name, reqError) {
         super(name.replace(/\n/g, ""), reqError);
         this.errors = [];