+++ CHANGELOG.md
@@ -4,6 +4,28 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [3.1.2] - 2025-03-12
+### ADDED
+- [CORE-1925](https://widesky.atlassian.net/browse/CORE-1925): Throw errors
+  `HaystackError` and `GraphQLError` now include the original request error
+  thrown by the client package in the `requestError` property, and the HTTP
+  status in `status` for convenience.
+- [CORE-1925](https://widesky.atlassian.net/browse/CORE-1925): Exported request
+  errors:
+  - `RequestError`: A base class for all request errors created.
+  - `HaystackError`: A error for a Haystack type response error.
+  - `GraphQLError`: A error for a GraphQL type response error.
+- [CORE-4036](https://widesky.atlassian.net/browse/CORE-4036): Added modules:
+  - `http2-wrapper`
+  - `axios-http2-adapter`
+- [CORE-4036](https://widesky.atlassian.net/browse/CORE-4036): `axios` module
+  now uses a `http2` adapter which allows it to make calls using the HTTP/2.0
+  protocol.
+
+### FIXED
+- [CORE-1907](https://widesky.atlassian.net/browse/CORE-1907): `HisWritePayload`
+  class not accepting boolean values.
+
 ## [3.1.1] - 2024-06-03
 ### CHANGED
 - The `http` option `keepAlive` will now default to `true` if it is not specified.